requirements.txt
cohere==4.1.4
discord.py==2.2.2
python-decouple==3.8
